fix: bug fix

See merge request sw_track/class_03/web_project/team5/team5-commerce!60

+++ src/services/product-service.js
@@ -1,24 +1,43 @@
-import { productModel } from '../db';
+import { productModel, categoryModel } from '../db';
 
 class ProductService {
-  constructor(productModel) {
+  constructor(productModel, categoryModel) {
     this.productModel = productModel;
+    this.categoryModel = categoryModel;
   }
 
   //상품 추가
   async addProduct(productInfo) {
+    const { title } = productInfo;
+    //상품 추가, 만약 같은 이름의 상품이 이미 있는 경우 추가하지 않음
+    console.log(title);
+    const product = await this.productModel.findByName(title);
+    console.log(product);
+    if (product) {
+      throw new Error(
+        '이 이름은 현재 사용중입니다. 다른 이름을 입력해 주세요..',
+      );
+    }
     const createdNewProduct = await this.productModel.create(productInfo);
     return createdNewProduct;
   }
 
   //상품 전체 목록을 받음
   async getProductlist() {
-    const products = await this.productModel.findAll({});
+    const products = await this.productModel.findAll();
     return products;
   }
 
   //해당 카테고리의 상품 전체 목록을 받음
   async getProductlistByCategory(categoryId) {
+    //해당 카테고리가 카테고리에 없는 경우 에러메시지 전달
+    console.log(categoryId);
+    const category = await this.categoryModel.findOneById(categoryId);
+    if (category) {
+      throw new Error(
+        '해당 id의 카테고리를 찾을 수 없습니다. 다시 확인해주세요.',
+      );
+    }
     const products = await this.productModel.findAllByCategory(categoryId);
     return products;
   }
@@ -30,6 +49,14 @@ class ProductService {
   }
   //상품 아이디로 가져옴
   async getProductByProductId(productId) {
+    // 우선 해당 id의 상품이 db에 있는지 확인
+    let foundProduct = await this.productModel.findById(productId);
+    // db에서 찾지 못한 경우, 에러 메시지 반환
+    if (!foundProduct) {
+      throw new Error(
+        '해당 productId의 상품이 없습니다. 다시 한 번 확인해 주세요.',
+      );
+    }
     const product = await this.productModel.findById(productId);
     return product;
   }
@@ -56,6 +83,14 @@ class ProductService {
 
   //상품 삭제
   async deleteProduct(productId) {
+    // 우선 해당 id의 상품이 db에 있는지 확인
+    let product = await this.productModel.findById(productId);
+    // db에서 찾지 못한 경우, 에러 메시지 반환
+    if (!product) {
+      throw new Error(
+        '해당 productId의 상품이 없습니다. 다시 한 번 확인해 주세요.',
+      );
+    }
     const deletedProduct = this.productModel.delete(productId);
     return deletedProduct;
   }
@@ -63,7 +98,7 @@ class ProductService {
   async getProductsByAdmin(query) {
     const [productTotal, products] = await Promise.all([
       this.productModel.getProductsCount(),
-      this.productModel.findAll(query),
+      this.productModel.findPage(query),
     ]);
     const { limit } = query;
     // const totalProductPage = Math.ceil(productTotal / limit);
@@ -72,6 +107,6 @@ class ProductService {
   }
 }
 
-const productService = new ProductService(productModel);
+const productService = new ProductService(productModel, categoryModel);
 
 export { productService };
